pyproject.toml
[project]
name = "langgraphproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.54.0",
    "ipython>=9.3.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.15",
    "langchain-community>=0.3.25",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.23",
    "langchain-tavily>=0.2.2",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
